miden_core_lib::dsa

Module eddsa_ed25519

EdDSA Ed25519 with SHA-512 signature helpers.

Functions in this module generate data for the miden::core::crypto::dsa::eddsa_ed25519::verify MASM procedure.

Functions

encode_signature

Encodes the provided public key and signature into a vector of field elements in the format expected by miden::core::crypto::dsa::eddsa_ed25519::verify procedure.

sign

Signs the provided message with the supplied secret key and encodes this signature and the associated public key into a vector of field elements in the format expected by miden::core::crypto::dsa::eddsa_ed25519::verify procedure.