miden_core_lib::dsa

Module falcon512_poseidon2

Falcon-512 with Poseidon2 hashing signature helpers.

Functions in this module generate data for the miden::core::crypto::dsa::falcon512_poseidon2::verify MASM procedure.

Structs

PublicKey

Public key represented as a polynomial with coefficients over the Falcon prime field.

SecretKey

Represents the secret key for Falcon DSA.

Signature

A deterministic Falcon512 Poseidon2 signature over a message.

Functions

encode_signature

Encodes the provided Falcon public key and signature into a vector of field elements in the format expected by miden::core::crypto::dsa::falcon512_poseidon2::verify procedure.

sign

Signs the provided message with the provided secret key and returns the resulting signature encoded in the format required by the falcon512_poseidon2::verify procedure, or None if the secret key is malformed due to either incorrect length or failed decoding.