miden_core_lib::handlers

Module sorted_array

Enums

SortedArrayError

Error types that can occur during LOWERBOUND event operations.

Constants

LOWERBOUND_ARRAY_EVENT_NAME

Event name for the lowerbound_array operation.

LOWERBOUND_KEY_VALUE_EVENT_NAME

Event name for the lowerbound_key_value operation.

Functions

handle_lowerbound_array

Pushes onto the advice stack the first pointer in [start_ptr, end_ptr) such that mem[word_ptr] >= KEY in lexicographic order of words. If all words are < KEY, returns end_ptr. The array must be sorted in non-decreasing order.

handle_lowerbound_key_value

Pushes onto the advice stack the first pointer in [start_ptr, end_ptr) such that mem[word_ptr] >= KEY in lexicographic order of words. If all keys are < KEY, returns end_ptr. The array must be a list of key-value pairs (word tuples) where all keys are in non-decreasing order.